% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{find_comments}
\alias{find_comments}
\title{Find a columns comments}
\usage{
find_comments(.data, ...)
}
\arguments{
\item{.data}{A data frame created by read_act}

\item{...}{One or more unquoted expressions seperated by commas.}
}
\value{
A data frame with a new column named commentText which has the text of the comments
}
\description{
Find a columns comments
}
\examples{
df <- readACT(filepath, scores_only = F)
df_with comments <- find_comments(df)
}
