% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{find_PTCRI}
\alias{find_PTCRI}
\title{Transform a column of PTCRI codes into text}
\usage{
find_PTCRI(.data, ...)
}
\arguments{
\item{.data}{A data frame created by read_ACT() with scores_only = F or another package function}

\item{...}{One or more unquoted expressions seperated by commas}
}
\value{
A data frame with a new column PTCRIText
df <- readACT(filepath, scores_only = F)
find_PTCRI(df)

df <- readACT(filepath, scores_only = F)
df <- find_comments(df)
df_with_PTCRI <- find_PTCRI(df)
}
\description{
Transform a column of PTCRI codes into text
}
