% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{find_relig_affil}
\alias{find_relig_affil}
\title{Transform a column of Religious Affiliation codes into text}
\usage{
find_relig_affil(.data, ...)
}
\arguments{
\item{.data}{A data frame created by read_ACT() with scores_only = F or another package function}

\item{...}{One or more unquoted expressions seperated by commas}
}
\value{
A data frame with a new column PTCRIText
df <- readACT(filepath, scores_only = F)
find_relig_affil(df)

df <- readACT(filepath, scores_only = F)
df <- find_comments(df)
df_with_relig_affil <- find_relig_affil(df)
}
\description{
Transform a column of Religious Affiliation codes into text
}
