% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{find_UCTI}
\alias{find_UCTI}
\title{Transform the UCTI code to level}
\usage{
find_UCTI(.data, ...)
}
\arguments{
\item{.data}{A data frame created by read_ACT() with scores_only = F or another package function}

\item{...}{One or more unquoted expressions seperated by commas}
}
\value{
A data frame with a new column UCTIText
}
\description{
Transform the UCTI code to level
}
\examples{
df <- readACT(filepath, scores_only = F)
find_UCTI(df)

df <- readACT(filepath, scores_only = F)
df <- find_comments(df)
df_with_UCTI <- find_UCTI(df)
}
