% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{get_comment}
\alias{get_comment}
\title{Translate comment code to plain text comment}
\usage{
get_comment(comment_code = "--")
}
\arguments{
\item{comment_code}{A two character code.}
}
\value{
A character with the text of the comment
}
\description{
Translate comment code to plain text comment
}
\examples{

get_comment("33")
get_comment("--")
}
