% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_file.R
\name{get_mapping}
\alias{get_mapping}
\title{Select the mapping to use when reading the ACT fixed width file}
\usage{
get_mapping(year = "19-20")
}
\arguments{
\item{year}{The school year for the mapping}
}
\value{
A data frame containing column names and widths
}
\description{
Select the mapping to use when reading the ACT fixed width file
}
