% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{get_PTCRI}
\alias{get_PTCRI}
\title{Progress Toward Career Readiness Indicator (PTCRI) code translation}
\usage{
get_PTCRI(textPTCRI = "-")
}
\arguments{
\item{textPTCRI}{The code for PTCRI level}
}
\value{
A text value of the indicated National Career Readiness certificate Projection.
}
\description{
Progress Toward Career Readiness Indicator (PTCRI) code translation
}
\examples{
get_PRCRI("4")
get_PRCRI("-")
}
