% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{get_relig_affil}
\alias{get_relig_affil}
\title{Religous affiliation code translation}
\usage{
get_relig_affil(textReligAffil = "NA")
}
\arguments{
\item{textReligAffil}{The code for religious affiliation}
}
\value{
A text value of the indicated religous affiliation
}
\description{
Religous affiliation code translation
}
