% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{get_UCTI}
\alias{get_UCTI}
\title{Convert Understanding Complex Text Indicator (UTCI) code to proficency level}
\usage{
get_UCTI(textUCTI = "-")
}
\arguments{
\item{textUCTI}{the text of a UTCI code}
}
\value{
The Proficency level of UTCI
}
\description{
Convert Understanding Complex Text Indicator (UTCI) code to proficency level
}
\examples{
get_UCTI("2")
}
