% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_file.R
\name{read_ACT}
\alias{read_ACT}
\title{Read in the ACT fixed width file}
\usage{
read_ACT(file = "fileName", year = "19-20", blanks = F, scores_only = T)
}
\arguments{
\item{file}{The file path to the file}

\item{year}{The school year for the mapping}

\item{blanks}{Boolean option to include the blank columns in the resulting data frame}

\item{scores_only}{Boolean option to only select scores.}
}
\value{
A data frame containing the data from the fixed width file provided by the ACT
}
\description{
Read in the ACT fixed width file
}
\examples{

#read_ACT(filepath, mapping = "19-20")
}
