% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_translations.R
\name{translate_comment}
\alias{translate_comment}
\title{Creates a string of comments based on the comment field}
\usage{
translate_comment(comment = "--------")
}
\arguments{
\item{comment}{An 8 character string consiting of 4 comment codes.}
}
\value{
A string of trasnlated comments.
}
\description{
Creates a string of comments based on the comment field
}
\examples{
translate_comment("010233--")
}
